# app/backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: BetterF API
  version: 0.1.1
  description:
    'BTF-5 implementation follows the product owner correction of 2026-10-04: separate account,
    organization, and email-delivery status records. Same-origin session cookie (HttpOnly, SameSite=Lax,
    Secure by default), 30-minute inactivity timeout. All mutations require CSRF.'
paths:
  /api/status:
    get:
      operationId: getStatus
      summary: Application and database availability
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusEnvelope'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/registration/roles:
    get:
      operationId: getProfessionalRoles
      summary: Predefined professional roles for onboarding
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesEnvelope'
  /api/registration:
    post:
      operationId: registerOrganization
      summary:
        Persist PENDING organization and account and send verification email. Repeated email submission
        reuses original data without overwriting credentials or sending again.
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingEnvelope'
        '400':
          description: Invalid required input or link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing/invalid CSRF or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'DUPLICATE_REGISTRATION: You cannot register this company. It is already registered.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description:
            Database or email unavailable; delivery failure preserves pending records and prior
            link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description:
            Unexpected failure; outcome may be uncertain. Retry through documented registration/resend
            behavior.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registration'
      parameters:
        - in: header
          name: X-CSRF-TOKEN
          required: true
          schema:
            type: string
          description: Masked token from /api/auth/csrf; include its session cookie. Refresh after login/logout.
  /api/registration/resend:
    post:
      operationId: resendVerification
      summary:
        Reuse email row; record attempt and SENT/FAILED status. Successful send rotates token for
        24h; failed send preserves old credential.
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingEnvelope'
        '400':
          description: Invalid required input or link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing/invalid CSRF or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'DUPLICATE_REGISTRATION: You cannot register this company. It is already registered.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description:
            Database or email unavailable; delivery failure preserves pending records and prior
            link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description:
            Unexpected failure; outcome may be uncertain. Retry through documented registration/resend
            behavior.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description:
            'RESEND_COOLDOWN: wait 60 seconds since the last successful send; current link
            unchanged'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
      parameters:
        - in: header
          name: X-CSRF-TOKEN
          required: true
          schema:
            type: string
          description: Masked token from /api/auth/csrf; include its session cookie. Refresh after login/logout.
  /api/registration/verify:
    post:
      operationId: verifyRegistration
      summary:
        Atomically activate pending account and organization and grant ADMIN. Recheck active-domain
        uniqueness; losing registrant gets no access.
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationEnvelope'
        '400':
          description: Invalid required input or link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing/invalid CSRF or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'DUPLICATE_REGISTRATION: You cannot register this company. It is already registered.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description:
            Database or email unavailable; delivery failure preserves pending records and prior
            link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description:
            Unexpected failure; outcome may be uncertain. Retry through documented registration/resend
            behavior.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - token
              properties:
                id:
                  type: string
                  format: uuid
                token:
                  type: string
                  maxLength: 100
                  writeOnly: true
      parameters:
        - in: header
          name: X-CSRF-TOKEN
          required: true
          schema:
            type: string
          description: Masked token from /api/auth/csrf; include its session cookie. Refresh after login/logout.
  /api/auth/csrf:
    get:
      operationId: getCsrf
      summary: Establish session and retrieve masked CSRF token. Call before mutations.
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfEnvelope'
  /api/auth/login:
    post:
      operationId: login
      summary:
        Email/password login; both account and organization must be ACTIVE. Rotate session ID and
        CSRF.
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountEnvelope'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 'LOGIN_FAILED: invalid credentials or inactive account/organization'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Invalid CSRF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  maxLength: 128
                  writeOnly: true
      parameters:
        - in: header
          name: X-CSRF-TOKEN
          required: true
          schema:
            type: string
          description: Masked token from /api/auth/csrf; include its session cookie. Refresh after login/logout.
  /api/auth/me:
    get:
      operationId: currentAccount
      summary: Read active authenticated account and organization; recheck both statuses
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountEnvelope'
        '403':
          description: Access denied or account/organization no longer active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/auth/logout:
    post:
      operationId: logout
      summary: Invalidate session and delete session cookie
      security:
        - SessionCookie: []
      parameters:
        - in: header
          name: X-CSRF-TOKEN
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Logged out, no body
        '403':
          description: Invalid CSRF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          type: 'null'
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    Role:
      type: object
      required:
        - id
        - label
      properties:
        id:
          type: string
        label:
          type: string
    Pending:
      type: object
      required:
        - email
        - resendAvailableAt
        - deliveryStatus
      properties:
        email:
          type: string
          format: email
        resendAvailableAt:
          type: string
          format: date-time
        deliveryStatus:
          type: string
          enum:
            - PENDING
            - SENT
            - FAILED
    Verification:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - VERIFIED
            - ALREADY_VERIFIED
    Csrf:
      type: object
      required:
        - token
        - headerName
      properties:
        token:
          type: string
        headerName:
          type: string
    Account:
      type: object
      required:
        - id
        - fullName
        - email
        - professionalRole
        - organizationId
        - organizationName
        - accessRole
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        professionalRole:
          type: string
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
        accessRole:
          type: string
          enum:
            - ADMIN
            - MEMBER
    Registration:
      type: object
      required:
        - companyName
        - website
        - specialization
        - fullName
        - email
        - password
        - professionalRole
      properties:
        companyName:
          type: string
          minLength: 1
          maxLength: 200
        website:
          type: string
          maxLength: 253
          description:
            HTTPS registry root domain; root slash optional. No subdomain, non-root path, query,
            fragment, credentials, or explicit port. Valid multi-label registry suffixes supported.
        specialization:
          type: string
          minLength: 1
          maxLength: 200
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 15
          maxLength: 128
          writeOnly: true
        professionalRole:
          type: string
          description: ID returned by GET /api/registration/roles. Unrelated to access role.
    PendingEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          $ref: '#/components/schemas/Pending'
        error:
          type: 'null'
    VerificationEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          $ref: '#/components/schemas/Verification'
        error:
          type: 'null'
    CsrfEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          $ref: '#/components/schemas/Csrf'
        error:
          type: 'null'
    AccountEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          $ref: '#/components/schemas/Account'
        error:
          type: 'null'
    RolesEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        error:
          type: 'null'
    StatusEnvelope:
      type: object
      required:
        - data
        - error
      properties:
        data:
          type: object
          required:
            - status
          properties:
            status:
              type: string
              enum:
                - UP
        error:
          type: 'null'